:ambulance: Fix a bug that caused files to be read incorrectly.
buf.slice().toString() turns out to be a bad idea. It completely
mangles most files. I didn't notice it in tests because the tests
use string inputs. Instead we are going one byte at a time, using
buf.readUInt8(). There is probably a better way to do this but
it will probably require exposing the node Buffer inside each
SourceBuffer (or, better yet, an overhaul of SourceBuffer that is
much more of a transparent wrapper around Buffer).